Add script for mintfor
